VELO-24 finished strava import endpoint

+++ src/main/kotlin/pl/edu/pw/mini/velobackend/domain/model/Location.kt
@@ -1,6 +1,6 @@
 package pl.edu.pw.mini.velobackend.domain.model
 
 data class Location(
-        val latitude: Double,
-        val longitude: Double
+        val latitude: Float,
+        val longitude: Float
 )
